backend/requirements.txt
Django~=3.2.4
django-cors-headers==3.7.0
django-filter~=2.4.0
djangorestframework~=3.12.4
drf-yasg==1.20.0
psycopg2==2.9.1
uvicorn==0.14.0
